BGR2010

Minnesota prior via dummy observations

BEAVARs.makeDummiesMinn! — Method

makeDummiesMinn!(sigma::Vector{Float64},delta,lambda,n::Integer,p::Integer,Y_d1))

Fills a matrix $Y_d$ and $X_d$ following eq. (5) in Banbura, Giannone, Reichling (2010), JAE, Large Bayesian Autoregressions.

\[Y_d = \begin{bmatrix} diag(σ_1*δ_1 \dots σ_N*δ_n)/λ\\ \mathbf{0}_{n*(p-1) \times n}\\ diag(σ_1, \dots, σ_n)\\ \mathbf{0}_{1,n \times n} \end{bmatrix}\\\]

\[X_d = \begin{bmatrix} diag(1,\dots,p) \otimes diag(σ_1, \dots σ_n)./ λ \quad \mathbf{0}_{n*p,1}\\ \mathbf{0}_{n,n*p+1}\\ \mathbf{0}_{1,n*p} ε\\ \end{bmatrix}\]

Instad of creating the matrix every time, the function uses mutation on matrices called Y_d1 and X_d1.

For Y_d1 it is the the diagonal of the first (1:n,1:n) block and in the diagonal of the third block n+n*(p-1)+1:n+n*(p-1)+n

For X_d1 it is the diagonal of the first (n * p x n * p), esesntially X_d1 is populated along its diagonal and only the constant is added at the end

function makeDummiesMinn!(sigma::Vector{Float64},delta::Vector{Float64},lambda,n::Integer,p::Integer,epsi,Y_d1,X_d1)
    CI1_Yd1 = CartesianIndex.(1:n,1:n)  # These are the diagonal indices for the top block 
    CI2_Yd1 = CartesianIndex.(n+n*(p-1)+1:n+n*(p-1)+n,1:n)  # These are the diagonal indices for the block in the middle
    CI_Xd1 = CartesianIndex.(1:n*p,1:n*p)
    Y_d1[CI1_Yd1] = sigma.*delta./lambda;
    Y_d1[CI2_Yd1] = sigma;
    X_d1[end,end] = epsi;
    X_d1[CI_Xd1]  = repeat(sigma./lambda,p).*repeat(1:p,inner=n)
    return Y_d1, X_d1;
end