CPZ2023

BEAVARs.CPZ2023 — Function

BEAVARs.CPZ2023(dataHF_tab::TimeArray{Typ,N,D,A},dataLF_tab::TimeArray{Typ,N,D,A},varOrder::Array{Symbol,1},varSetup::BVARmodelSetup,hyp_struct::BVARmodelHypSetup)

Estimate Chan, Zhu, Poon 2024 using a  Minnesota-based independent Normal-Wishart prior

BEAVARs.makeDataSetup — Method

makeDataSetup(::CPZ2023_type,dataHF_tab::TimeArray, dataLF_tab::TimeArray; var_list =  [colnames(dataHF_tab); colnames(dataLF_tab)])

Generate data for a mixed-frequency VAR. Uses Time Arrays from the TimeSeries package

Arguments

dataHF_tab: TimeArray with your high-frequency variables (monthly or quarterly, respectively)
dataLF_tab: TimeArray with your low-frequency variables (quarterly or yearly, respectively)
var_list:   the variable order. Note that the functions that call these variables allow this to be optional.

See also dataCPZ2023.

BEAVARs.CPZ_prep_TimeArrays — Method

CPZ_prep_TimeArrays(dataLF_tab::TimeArray{T,N,D,A},dataHF_tab::TimeArray{T,N,D,A},varOrder::Array{Symbol,1},prior_RW::Int,n_fcst::Int)

Prepare one large table with the full dataset  `fataHF_tab` with both low and high-frequency variables with low-freq having `NaNs`. 
It will extend the table with n_fcst, which is the amoung of *low frequency* time periods.

varOrder must be a `Vector{Symbol}` and not `Vector{Vector{Symbol}}`
e.g. [varNamesLF; varNamesHF] and not [varNamesLF, varNamesHF]
prior_RW = 0: growth rates, 1: log-levels. indicator for the aggregate weights in the inter-temporal aggregation

BEAVARs.CPZ_makeM_inter — Method

CPZ_makeM_inter(z_tab,YYt,Sm_bit,datesHF,varNamesLF,fvarNames,freq_mix_tp,nm,Tf;scVal=10e-8)

Generates an M matrix such that z = My + eps. y is a vector of unobserved high-frequency values and the observations in z relate by a linear combination to a subset of the values in y

Arguments

z_tab:      A `TimeArray` of low-frequncy observations, either yearly or quarterly
YYt:        A transposed vector of data. Only its dimensions matter, not what is in it. It is assumed it has some unobserved values which are observed only at the lower frequency (`z_tab`)
Sm_bit:     A boolean matrix indicating which indices in YYt are to be matched, i.e. which are the missing high-frequency observations
datesHF:    A TimeArray accompanying `YYt`
varNamesLF: Vector with variable names from `z_tab` to be matched with a the high-frequency data. Code will err if varNamesLF is not a subset of fvarNames
fvarNames:  Vector with variable names from `y` to be matched with a the low-frequency data. Code will err if varNamesLF is not a subset of fvarNames
freq_mix_tp:Tuple indicating what is the frequency mix: monthly, quarterly, yearly, growth rates, log-levels
nm:         Total number of missing values  (consider dropping this, because we have Sm_bit)
Tf:         Total number of high-frequency observations (Tf = length(YY), Tf = size(YYt,2), consider dropping this as input, as we have YYt in the function)
scVal:      Optional parameter, value of how "hard" the constraint is. Setting it to very small values enforces z = My, while setting it higher allows for some discrepancies

Important

The function currently accepts only one variable in ztab, has to be extended. If you have both GDP and Consumption which are observed in LF and Ztab has 2 columns, the O-matrix will be wrong, see below

TODO: add support for z_tab variables with various lengths

BEAVARs.CPZ_makeM_inter_agg — Method

M_inter_agg = CPZ_makeM_inter_agg(fdatesLF,fdatesHF,freq_mix_tp)

Generate an M matrix for aggregating vectors of low frequency to high frequency data such that M*high_freq_tab = low_freq_tab

BEAVARs.CPZ_initMatrices — Method

BEAVARs.CPZ_draw! — Method

BEAVARs.CPZ_draw_wz! — Method

Draw with restrictions

BEAVARs.CPZ_iniw! — Method

Updates parameters using an independennt Normal-Wishart prior

Updates (i.e. overwrites): 
    β_draw, 
    B_draw, 
    structB_draw,
    Σt,
    Σt_inv

BEAVARs.CPZ_init_Minn — Method

Y, X, T, deltaP, sigmaP, mu_prior, Vβ_inv, Vβ_inv_diag_view, XtΣ_inv_den, XtΣ_inv_X, Xsur_den, Xsur_CI, X_CI, k, intercept, K_β, beta,  = CPZ_init_Minn(YY,p)

Initializes matrices for using the Minnesota prior in the CPZ2023 framework

link to References